+++ test/doctype.test.js
@@ -1,9 +1,10 @@
 const { here } = require("./utils");
 
 describe("doctype", () => {
-  test("prints out first line if it's a doctype", () => {
+  test("renders appropriately", () => {
     const content = here(`
-      <?xml version="1.0" encoding="UTF-8"?>
+      <?xml version="1.0" encoding="UTF-8" ?>
+      <?xml-model href="model.rnc" type="application/relax-ng-compact-sync-syntax" ?>
       <foo />
     `);