+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - Support for cdata tags.
 
+### Changed
+
+- Now using `parser.getTraversalObject` instead of `parser.parse`. This way there's less transformation and it leaves the option open to eventually just use the parser directly.
+
 ## [0.2.0] - 2019-11-12
 
 ### Changed