+++ CHANGELOG.md
@@ -9,10 +9,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 ### Added
 
 - Support for cdata tags.
+- Support for the `locStart` and `locEnd` functions by tracking node metadata in the new parser.
+- Support for comment nodes.
+- Support for `<?xml ... ?>` and `<?xml-model ... ?>` tags.
 
 ### Changed
 
-- Now using `parser.getTraversalObject` instead of `parser.parse`. This way there's less transformation and it leaves the option open to eventually just use the parser directly.
+- Dropped the dependency on `fast-xml-parser` in favor of writing our own for better control over comments and node location.
 
 ## [0.2.0] - 2019-11-12